Try to delete file if not appending and is not writable.

+++ modules/apb-base/src/apb/utils/FileUtils.java
@@ -428,6 +428,11 @@ public class FileUtils
 
         parentFile.mkdirs();
 
+        if (!append && file.exists() && !file.canWrite() && !file.delete()) {
+            throw new BuildException("Can not recreate: '" + file + "'.");
+        }
+
+
         try {
             return new FileOutputStream(file, append);
         }
